refactoring : replaced handleFillGuestData logic into useLogInForm hook.
